requirements.txt
--index-url https://download.pytorch.org/whl/cu118

torch==2.0.0

scikit-learn
numpy

pyyaml
tqdm